--- test.txt
@@ -1,1 +1,0 @@
-11111111111111111111111111
